requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
reportlab==4.0.8
pypdf2==3.0.1
python-dotenv==1.0.0
openai==1.3.5
python-dateutil==2.8.2
pytest==7.4.3
black==23.11.0
pylint==3.0.2
streamlit==1.29.0
streamlit-extras==0.3.6